+++ packages/iov-tendermint-rpc/src/client.spec.ts
@@ -1,9 +1,10 @@
 // tslint:disable:no-console readonly-array
 import { ReadonlyDate } from "readonly-date";
+import { Stream } from "xstream";
 
 import { Encoding } from "@iov/encoding";
 
-import { Adaptor, v0_25 } from "./adaptor";
+import { Adaptor, adatorForVersion } from "./adaptor";
 import { Client } from "./client";
 import { randomId } from "./common";
 import { buildTagsQuery, QueryTag } from "./requests";
@@ -20,41 +21,57 @@ function pendingWithoutTendermint(): void {
   }
 }
 
-// TODO: make flexible, support multiple versions, etc...
-const tendermintUrl = "localhost:12345";
+/**
+ * Tendermint instances to be tested.
+ *
+ * Testing legacy version: as a convention, the minor version number is encoded
+ * in the port 111<version>, e.g. Tendermint 0.21.0 runs on port 11121. To start
+ * a legacy version use
+ *   TENDERMINT_VERSION=0.21.0 TENDERMINT_PORT=11121 ./scripts/tendermint/start.sh
+ *
+ * When more than 1 instances of tendermint are running, stop them manually:
+ *   docker container ls | grep tendermint/tendermint
+ *   docker container kill <container id from 1st column>
+ */
+const tendermintInstances = [
+  // {
+  //   url: "localhost:11121",
+  //   version: "0.21.x",
+  // },
+  {
+    url: "localhost:12345",
+    version: "0.25.x",
+  },
+];
+
+function sleep(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
 
 function buildKvTx(k: string, v: string): Uint8Array {
   return Encoding.toAscii(`${k}=${v}`);
 }
 
-function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
-  const key = randomId();
-  const value = randomId();
-
-  it("Tries to connect with known version to tendermint", async () => {
+function defaultTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
+  it("can connect to tendermint with known version", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
     expect(await client.abciInfo()).toBeTruthy();
+    client.disconnect();
   });
 
-  it("Tries to auto-discover tendermint", async () => {
+  it("can auto-discover tendermint version and connect", async () => {
     pendingWithoutTendermint();
     const client = await Client.detectVersion(rpcFactory());
     const info = await client.abciInfo();
     expect(info).toBeTruthy();
-  });
-
-  it("can disconnect", async () => {
-    pendingWithoutTendermint();
-    const client = await Client.detectVersion(rpcFactory());
-    await client.abciInfo();
     client.disconnect();
   });
 
-  it("Posts a transaction", async () => {
+  it("can post a transaction", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
-    const tx = buildKvTx(key, value);
+    const tx = buildKvTx(randomId(), randomId());
 
     const response = await client.broadcastTxCommit({ tx: tx });
     expect(response.height).toBeGreaterThan(2);
@@ -65,23 +82,30 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     if (response.deliverTx) {
       expect(response.deliverTx.code).toBeFalsy();
     }
+
+    client.disconnect();
   });
 
-  it("Queries the state", async () => {
+  it("can query the state", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
+
+    const key = randomId();
+    const value = randomId();
+    await client.broadcastTxCommit({ tx: buildKvTx(key, value) });
 
     const binKey = Encoding.toAscii(key);
     const binValue = Encoding.toAscii(value);
     const queryParams = { path: "/key", data: binKey };
-
     const response = await client.abciQuery(queryParams);
-    expect(new Uint8Array(response.key)).toEqual(binKey);
-    expect(new Uint8Array(response.value)).toEqual(binValue);
+    expect(response.key).toEqual(binKey);
+    expect(response.value).toEqual(binValue);
     expect(response.code).toBeFalsy();
+
+    client.disconnect();
   });
 
-  it("Sanity check - calls don't error", async () => {
+  it("can call a bunch of methods", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
 
@@ -93,9 +117,11 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     expect(await client.health()).toBeNull();
     expect(await client.status()).toBeTruthy();
     expect(await client.validators()).toBeTruthy();
+
+    client.disconnect();
   });
 
-  it("Can query a tx properly", async () => {
+  it("can query a tx properly", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
 
@@ -104,7 +130,7 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     const tx = buildKvTx(find, me);
 
     const txRes = await client.broadcastTxCommit({ tx });
-    expect(responses.txCommitSuccess(txRes)).toBeTruthy();
+    expect(responses.broadcastTxCommitSuccess(txRes)).toEqual(true);
     expect(txRes.height).toBeTruthy();
     const height: number = txRes.height || 0; // || 0 for type system
     expect(txRes.hash.length).not.toEqual(0);
@@ -115,7 +141,7 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     // both values come from rpc, so same type (Buffer/Uint8Array)
     expect(r.hash).toEqual(hash);
     // force the type when comparing to locally generated value
-    expect(new Uint8Array(r.tx)).toEqual(tx);
+    expect(r.tx).toEqual(tx);
     expect(r.height).toEqual(height);
     expect(r.proof).toBeTruthy();
 
@@ -143,9 +169,11 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     expect(block.blockMeta.header.numTxs).toEqual(1);
     expect(block.block.txs.length).toEqual(1);
     expect(block.block.txs[0]).toEqual(tx);
+
+    client.disconnect();
   });
 
-  it("Can paginate over all txs", async () => {
+  it("can paginate over txSearch results", async () => {
     pendingWithoutTendermint();
     const client = new Client(rpcFactory(), adaptor);
 
@@ -157,7 +185,7 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
       const tx = buildKvTx(find, me);
 
       const txRes = await client.broadcastTxCommit({ tx });
-      expect(responses.txCommitSuccess(txRes)).toBeTruthy();
+      expect(responses.broadcastTxCommitSuccess(txRes)).toEqual(true);
       expect(txRes.height).toBeTruthy();
       expect(txRes.hash.length).not.toEqual(0);
     };
@@ -166,6 +194,8 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     await sendTx();
     await sendTx();
     await sendTx();
+
+    await sleep(50); // Tendermint needs some time to update search index
 
     // expect one page of results
     const s1 = await client.txSearch({ query, page: 1, per_page: 2 });
@@ -185,6 +215,8 @@ function kvTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void {
     const [tx1, tx2, tx3] = sall.txs;
     expect(tx2.height).toEqual(tx1.height + 1);
     expect(tx3.height).toEqual(tx2.height + 1);
+
+    client.disconnect();
   });
 }
 
@@ -205,7 +237,8 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
           expect(event.height).toBeGreaterThan(0);
           // seems that tendermint just guarantees within the last second for timestamp
           expect(event.time.getTime()).toBeGreaterThan(testStart - 1000);
-          expect(event.time.getTime()).toBeLessThanOrEqual(ReadonlyDate.now());
+          // Tendermint clock is sometimes ahead of test clock. Add 10ms tolerance
+          expect(event.time.getTime()).toBeLessThanOrEqual(ReadonlyDate.now() + 10);
           expect(event.numTxs).toEqual(0);
           expect(event.lastBlockId).toBeTruthy();
           expect(event.totalTxs).toBeGreaterThan(0);
@@ -236,13 +269,15 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
             expect(events[1].lastCommitHash).not.toEqual(events[0].lastCommitHash);
             expect(events[1].lastResultsHash).not.toEqual(events[0].lastResultsHash);
             expect(events[1].validatorsHash).toEqual(events[0].validatorsHash);
+
+            client.disconnect();
             done();
           }
         },
-        error: fail,
-        complete: () => fail("Stream must not close just because we don't listen anymore"),
+        error: done.fail,
+        complete: () => done.fail("Stream completed before we are done"),
       });
-    })().catch(fail);
+    })().catch(done.fail);
   });
 
   it("can subscribe to block events", done => {
@@ -261,7 +296,8 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
           expect(event.header.height).toBeGreaterThan(0);
           // seems that tendermint just guarantees within the last second for timestamp
           expect(event.header.time.getTime()).toBeGreaterThan(testStart - 1000);
-          expect(event.header.time.getTime()).toBeLessThanOrEqual(ReadonlyDate.now());
+          // Tendermint clock is sometimes ahead of test clock. Add 10ms tolerance
+          expect(event.header.time.getTime()).toBeLessThanOrEqual(ReadonlyDate.now() + 10);
           expect(event.header.numTxs).toEqual(1);
           expect(event.header.lastBlockId).toBeTruthy();
           expect(event.header.totalTxs).toBeGreaterThan(0);
@@ -287,11 +323,13 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
 
             expect(events[1].header.appHash).not.toEqual(events[0].header.appHash);
             expect(events[1].header.validatorsHash).toEqual(events[0].header.validatorsHash);
+
+            client.disconnect();
             done();
           }
         },
-        error: fail,
-        complete: () => fail("Stream must not close just because we don't listen anymore"),
+        error: done.fail,
+        complete: () => done.fail("Stream completed before we are done"),
       });
 
       const transaction1 = buildKvTx(randomId(), randomId());
@@ -299,7 +337,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
 
       await client.broadcastTxCommit({ tx: transaction1 });
       await client.broadcastTxCommit({ tx: transaction2 });
-    })().catch(fail);
+    })().catch(done.fail);
   });
 
   it("can subscribe to transaction events", done => {
@@ -324,11 +362,13 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
             expect(events.length).toEqual(2);
             expect(events[1].height).toEqual(events[0].height + 1);
             expect(events[1].result.tags).not.toEqual(events[0].result.tags);
+
+            client.disconnect();
             done();
           }
         },
-        error: fail,
-        complete: () => fail("Stream must not close just because we don't listen anymore"),
+        error: done.fail,
+        complete: () => done.fail("Stream completed before we are done"),
       });
 
       const transaction1 = buildKvTx(randomId(), randomId());
@@ -336,7 +376,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
 
       await client.broadcastTxCommit({ tx: transaction1 });
       await client.broadcastTxCommit({ tx: transaction2 });
-    })().catch(fail);
+    })().catch(done.fail);
   });
 
   it("can subscribe to transaction events filtered by creator", done => {
@@ -362,11 +402,13 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
             expect(events.length).toEqual(2);
             expect(events[1].height).toEqual(events[0].height + 1);
             expect(events[1].result.tags).not.toEqual(events[0].result.tags);
+
+            client.disconnect();
             done();
           }
         },
-        error: fail,
-        complete: () => fail("Stream must not close just because we don't listen anymore"),
+        error: done.fail,
+        complete: () => done.fail("Stream completed before we are done"),
       });
 
       const transaction1 = buildKvTx(randomId(), randomId());
@@ -374,39 +416,83 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor): void
 
       await client.broadcastTxCommit({ tx: transaction1 });
       await client.broadcastTxCommit({ tx: transaction2 });
-    })().catch(fail);
+    })().catch(done.fail);
+  });
+
+  // This a a minimal showcase for subscribing to the same query twice
+  // See https://github.com/iov-one/iov-core/issues/581
+  xit("can subscribe twice", done => {
+    pendingWithoutTendermint();
+
+    (async () => {
+      const events: responses.NewBlockHeaderEvent[] = [];
+      const client = new Client(rpcFactory(), adaptor);
+      const stream1 = client.subscribeNewBlockHeader();
+      const stream2 = client.subscribeNewBlockHeader();
+
+      const subscription = Stream.merge(stream1, stream2).subscribe({
+        next: event => {
+          events.push(event);
+
+          // collect 2x2 events
+          if (events.length === 4) {
+            // two events per height
+            expect(new Set(events.map(e => e.height)).size).toEqual(2);
+
+            subscription.unsubscribe();
+            client.disconnect();
+            done();
+          }
+        },
+        error: done.fail,
+        complete: () => done.fail("Stream completed before we are done"),
+      });
+    })().catch(done.fail);
   });
 }
 
-describe("Client", () => {
-  it("can connect to a given url", async () => {
-    pendingWithoutTendermint();
+for (const { url, version } of tendermintInstances) {
+  describe(`Client ${version}`, () => {
+    it("can connect to a given url", async () => {
+      pendingWithoutTendermint();
 
-    // default connection
-    const client = await Client.connect(tendermintUrl);
-    const info = await client.abciInfo();
-    expect(info).toBeTruthy();
+      // default connection
+      {
+        const client = await Client.connect(url);
+        const info = await client.abciInfo();
+        expect(info).toBeTruthy();
+        client.disconnect();
+      }
 
-    // http connection
-    const client2 = await Client.connect("http://" + tendermintUrl);
-    const info2 = await client2.abciInfo();
-    expect(info2).toBeTruthy();
+      // http connection
+      {
+        const client = await Client.connect("http://" + url);
+        const info = await client.abciInfo();
+        expect(info).toBeTruthy();
+        client.disconnect();
+      }
 
-    // ws connection
-    const client3 = await Client.connect("ws://" + tendermintUrl);
-    const info3 = await client3.abciInfo();
-    expect(info3).toBeTruthy();
+      // ws connection
+      {
+        const client = await Client.connect("ws://" + url);
+        const info = await client.abciInfo();
+        expect(info).toBeTruthy();
+        client.disconnect();
+      }
+    });
+
+    describe("With HttpClient", () => {
+      const adaptor = adatorForVersion(version);
+      defaultTestSuite(() => new HttpClient(url), adaptor);
+    });
+
+    describe("With WebsocketClient", () => {
+      // don't print out WebSocket errors if marked pending
+      const onError = skipTests() ? () => 0 : console.log;
+      const factory = () => new WebsocketClient(url, onError);
+      const adaptor = adatorForVersion(version);
+      defaultTestSuite(factory, adaptor);
+      websocketTestSuite(factory, adaptor);
+    });
   });
-
-  describe("With HttpClient: v0-25", () => {
-    kvTestSuite(() => new HttpClient(tendermintUrl), v0_25);
-  });
-
-  describe("With WebsocketClient: v0-25", () => {
-    // don't print out WebSocket errors if marked pending
-    const onError = skipTests() ? () => 0 : console.log;
-    const factory = () => new WebsocketClient(tendermintUrl, onError);
-    kvTestSuite(factory, v0_25);
-    websocketTestSuite(factory, v0_25);
-  });
-});
+}