+++ packages/iov-tendermint-rpc/src/requests.ts
@@ -2,25 +2,29 @@ import { As } from "type-tagger";
 
 import { JsonRpcRequest, jsonRpcWith } from "./common";
 
-// union type of all possible methods?
+/**
+ * RPC methods as documented in https://tendermint.com/rpc/
+ *
+ * Enum raw value must match the spelling in the "shell" example call (snake_case)
+ */
 export const enum Method {
-  ABCI_INFO = "abci_info",
-  ABCI_QUERY = "abci_query",
-  BLOCK = "block",
-  BLOCKCHAIN = "blockchain",
-  BLOCK_RESULTS = "block_results",
-  BROADCAST_TX_ASYNC = "broadcast_tx_async",
-  BROADCAST_TX_SYNC = "broadcast_tx_sync",
-  BROADCAST_TX_COMMIT = "broadcast_tx_commit",
-  COMMIT = "commit",
-  GENESIS = "genesis",
-  HEALTH = "health",
-  STATUS = "status",
-  SUBSCRIBE = "subscribe",
-  TX = "tx",
-  TX_SEARCH = "tx_search",
-  VALIDATORS = "validators",
-  UNSUBSCRIBE = "unsubscribe",
+  AbciInfo = "abci_info",
+  AbciQuery = "abci_query",
+  Block = "block",
+  Blockchain = "blockchain",
+  BlockResults = "block_results",
+  BroadcastTxAsync = "broadcast_tx_async",
+  BroadcastTxSync = "broadcast_tx_sync",
+  BroadcastTxCommit = "broadcast_tx_commit",
+  Commit = "commit",
+  Genesis = "genesis",
+  Health = "health",
+  Status = "status",
+  Subscribe = "subscribe",
+  Tx = "tx",
+  TxSearch = "tx_search",
+  Validators = "validators",
+  Unsubscribe = "unsubscribe",
 }
 
 export type Request =
@@ -50,11 +54,11 @@ export enum SubscriptionEventType {
 }
 
 export interface AbciInfoRequest {
-  readonly method: Method.ABCI_INFO;
+  readonly method: Method.AbciInfo;
 }
 
 export interface AbciQueryRequest {
-  readonly method: Method.ABCI_QUERY;
+  readonly method: Method.AbciQuery;
   readonly params: AbciQueryParams;
 }
 export interface AbciQueryParams {
@@ -65,14 +69,14 @@ export interface AbciQueryParams {
 }
 
 export interface BlockRequest {
-  readonly method: Method.BLOCK;
+  readonly method: Method.Block;
   readonly params: {
     readonly height?: number;
   };
 }
 
 export interface BlockchainRequest {
-  readonly method: Method.BLOCKCHAIN;
+  readonly method: Method.Blockchain;
   readonly params: BlockchainRequestParams;
 }
 export interface BlockchainRequestParams {
@@ -81,14 +85,14 @@ export interface BlockchainRequestParams {
 }
 
 export interface BlockResultsRequest {
-  readonly method: Method.BLOCK_RESULTS;
+  readonly method: Method.BlockResults;
   readonly params: {
     readonly height?: number;
   };
 }
 
 export interface BroadcastTxRequest {
-  readonly method: Method.BROADCAST_TX_ASYNC | Method.BROADCAST_TX_SYNC | Method.BROADCAST_TX_COMMIT;
+  readonly method: Method.BroadcastTxAsync | Method.BroadcastTxSync | Method.BroadcastTxCommit;
   readonly params: BroadcastTxParams;
 }
 export interface BroadcastTxParams {
@@ -96,26 +100,26 @@ export interface BroadcastTxParams {
 }
 
 export interface CommitRequest {
-  readonly method: Method.COMMIT;
+  readonly method: Method.Commit;
   readonly params: {
     readonly height?: number;
   };
 }
 
 export interface GenesisRequest {
-  readonly method: Method.GENESIS;
+  readonly method: Method.Genesis;
 }
 
 export interface HealthRequest {
-  readonly method: Method.HEALTH;
+  readonly method: Method.Health;
 }
 
 export interface StatusRequest {
-  readonly method: Method.STATUS;
+  readonly method: Method.Status;
 }
 
 export interface SubscribeRequest {
-  readonly method: Method.SUBSCRIBE;
+  readonly method: Method.Subscribe;
   readonly query: {
     readonly type: SubscriptionEventType;
     readonly tags?: ReadonlyArray<QueryTag>;
@@ -130,7 +134,7 @@ export interface QueryTag {
 }
 
 export interface TxRequest {
-  readonly method: Method.TX;
+  readonly method: Method.Tx;
   readonly params: TxParams;
 }
 export interface TxParams {
@@ -140,7 +144,7 @@ export interface TxParams {
 
 // TODO: clarify this type
 export interface TxSearchRequest {
-  readonly method: Method.TX_SEARCH;
+  readonly method: Method.TxSearch;
   readonly params: TxSearchParams;
 }
 export interface TxSearchParams {
@@ -151,7 +155,7 @@ export interface TxSearchParams {
 }
 
 export interface ValidatorsRequest {
-  readonly method: Method.VALIDATORS;
+  readonly method: Method.Validators;
   readonly params: {
     readonly height?: number;
   };